Fix NPE in Android{S,N}dkRepositoryFunction.
If neither the path attribute or the $ANDROID_HOME/$ANDROID_NDK_HOME environment
variable is set, we should report a helpful error message. A previous change
switched the lookup for the environment variable from the client environment map
to the skyframe environment map. However in the skyframe action environment map,
environment variables that are requested but not set in the environment have
null values.

Followup to change needed for #2621.

--
PiperOrigin-RevId: 149369076
MOS_MIGRATED_REVID=149369076